requirements.txt
asgiref==3.11.1
dj-database-url==2.3.0
Django==5.2.11
django-autoslug==1.9.9
django-ckeditor==6.7.3
django-environ==0.13.0
django-js-asset==3.1.2
gunicorn==22.0.0
packaging==26.0
pillow==10.4.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
